iap_social/requirements_social.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
replicate>=0.25.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
notion-client>=2.2.1
sentence-transformers>=2.2.0
